Ajuste de tipo de parametros

+++ AllInMobileLibrary/pushnotifications-library/src/main/java/br/com/allin/mobile/pushnotification/service/allin/MessageService.java
@@ -3,7 +3,7 @@ package br.com.allin.mobile.pushnotification.service.allin;
 import br.com.allin.mobile.pushnotification.task.allin.MessageTask;
 
 public class MessageService {
-    public static void markAsRead(long notificationId) {
+    public static void markAsRead(int notificationId) {
         new MessageTask(notificationId).execute();
     }
 }
